requirements.txt
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
xgboost==2.0.3
websockets==12.0
fastapi==0.109.2
uvicorn==0.27.0.post1
python-dotenv==1.0.1
requests==2.31.0
